ref: moved things around, feat: some things now working ;)
