+++ Cargo.toml
@@ -21,3 +21,4 @@ tokio = { version = "1.43.0", features = ["full"] }
 async-trait = "0.1.86"
 futures = "0.3.31"
 clap = { version = "4.5.31", features = ["derive"] }
+tokio-test = "0.4.4"
