+++ src/main.rs
@@ -3,6 +3,11 @@ pub mod conf;
 pub mod pkg;
 pub mod prelude;
 
-fn main() {
+use prelude::Result;
+
+#[tokio::main]
+async fn main() -> Result<()> {
     tracing_subscriber::fmt::init();
+    cmd::run().await?;
+    Ok(())
 }
