+++ src/pkg/server/loader.rs
@@ -12,7 +12,7 @@ use crate::pkg::conf::spec::{Config, Spec};
 use super::Server;
 
 impl Server {
-    fn new() -> Result<Server> {
+    pub fn new() -> Result<Server> {
         let config_path =
             env::var("LITEGINX_CONF_DIR").unwrap_or(format!("{}/.config/liteginx", env!("HOME")));
         let configs: Vec<Config> = fs::read_dir(&config_path)?
